requirements/api.txt
-r base.txt
djangorestframework==3.1.1
djangorestframework-gis==0.8.2
